Move runbit logic to appropriate subpkg
