+++ internal/runbits/runtime/runtimeevents.go
@@ -1,4 +1,4 @@
-package runbits
+package runtime
 
 import (
 	"io"
