+++ internal/scriptrun/scriptrun.go
@@ -16,7 +16,7 @@ import (
 	"github.com/ActiveState/cli/internal/output"
 	"github.com/ActiveState/cli/internal/process"
 	"github.com/ActiveState/cli/internal/rtutils"
-	"github.com/ActiveState/cli/internal/runbits"
+	rtrunbit "github.com/ActiveState/cli/internal/runbits/runtime"
 	"github.com/ActiveState/cli/internal/scriptfile"
 	"github.com/ActiveState/cli/internal/subshell"
 	"github.com/ActiveState/cli/internal/virtualenvironment"
@@ -73,7 +73,7 @@ func (s *ScriptRun) PrepareVirtualEnv() (rerr error) {
 	case err == nil:
 		break
 	case errors.Is(err, runtime.NeedsUpdateError):
-		pg := runbits.NewRuntimeProgressIndicator(s.out)
+		pg := rtrunbit.NewRuntimeProgressIndicator(s.out)
 		defer rtutils.Closer(pg.Close, &rerr)
 		if err := rt.Update(pg); err != nil {
 			return locale.WrapError(err, "err_update_runtime", "Could not update runtime installation.")
@@ -85,7 +85,7 @@ func (s *ScriptRun) PrepareVirtualEnv() (rerr error) {
 	}
 
 	if rt.NeedsUpdate() {
-		pg := runbits.NewRuntimeProgressIndicator(s.out)
+		pg := rtrunbit.NewRuntimeProgressIndicator(s.out)
 		defer rtutils.Closer(pg.Close, &rerr)
 		if err := rt.SolveAndUpdate(pg); err != nil {
 			return locale.WrapError(err, "err_update_runtime", "Could not update runtime installation.")
